【柴方晨】修改条件发券逻辑

+++ card-coupon/src/main/java/com/joycity/joyclub/card_coupon/service/sendcardstrategy/SingleStrategy.java
@@ -1,6 +1,8 @@
 package com.joycity.joyclub.card_coupon.service.sendcardstrategy;
 
+import java.math.BigDecimal;
+
 public interface SingleStrategy {
 
-    public Integer CouponNum(Float amount, Float criteriaAmount);
+    public Integer CouponNum(BigDecimal amount, BigDecimal criteriaAmount);
 }
